requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
openpyxl==3.1.2
pandas==2.2.0
python-dotenv==1.0.1
requests==2.31.0
